Move `upgrade-elements` to the `base/` dir
This module is not a controller itself but a helper function used to
register and instantiate controllers for elements on the page.
