+++ h/static/scripts/tests/base/upgrade-elements-test.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var upgradeElements = require('../../controllers/upgrade-elements');
+var upgradeElements = require('../../base/upgrade-elements');
 
 describe('upgradeElements', function () {
   it('should upgrade elements matching selectors', function () {
